+++ app/src/main/res/layout/activity_repo_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".RepoListActivity">
 
     <TextView
         android:layout_width="wrap_content"
